Refactor projects/remote API into separate modules
